Added functionality of login form choice,test authentication and help for plugin configuration and setup
git-svn-id: https://plugins.svn.wordpress.org/miniorange-2-factor-authentication/trunk@1213588 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ miniorange_2_factor_configuration.php
@@ -4,38 +4,58 @@
 			<p style="color:red;">(Warning: <a href="http://php.net/manual/en/curl.installation.php" target="_blank">PHP CURL extension</a> is not installed or disabled)</p>
 		<?php
 		}
+		
+		$mo2f_active_tab = isset($_GET['mo2f_tab']) ? $_GET['mo2f_tab'] : '2factor_setup';
 		?>
+		<div id="tab">
+			<h2 class="nav-tab-wrapper">
+				<a href="admin.php?page=miniOrange_2_factor_settings&amp;mo2f_tab=2factor_setup" class="nav-tab <?php echo $mo2f_active_tab == '2factor_setup' ? 'nav-tab-active' : ''; ?>" id="mo2f_tab1">
+				<?php if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_INITIALIZE_MOBILE_REGISTRATION' || get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_PLUGIN_SETTINGS'){ ?>User Profile <?php }else{ ?> Account Setup <?php } ?></a> 
+				<a href="admin.php?page=miniOrange_2_factor_settings&amp;mo2f_tab=mobile_configure" class="nav-tab <?php echo $mo2f_active_tab == 'mobile_configure' ? 'nav-tab-active' : ''; ?>" id="mo2f_tab3">Configure Mobile</a>
+				<a href="admin.php?page=miniOrange_2_factor_settings&amp;mo2f_tab=mo2f_login" class="nav-tab <?php echo $mo2f_active_tab == 'mo2f_login' ? 'nav-tab-active' : ''; ?>" id="mo2f_tab2">Login Settings</a>
+			    <a href="admin.php?page=miniOrange_2_factor_settings&amp;mo2f_tab=mo2f_demo" class="nav-tab <?php echo $mo2f_active_tab == 'mo2f_demo' ? 'nav-tab-active' : ''; ?>" id="mo2f_tab4">How It Works</a>
+			    <a href="admin.php?page=miniOrange_2_factor_settings&amp;mo2f_tab=mo2f_help" class="nav-tab <?php echo $mo2f_active_tab == 'mo2f_help' ? 'nav-tab-active' : ''; ?>" id="mo2f_tab5">Help & Troubleshooting</a>
+			</h2>
+		</div>
 		<div class="mo2f_container">
 			<table style="width:100%;">
 				<tr>
 					<td style="width:60%;vertical-align:top;">
 						<?php
 							
-							if(get_option( 'mo_2factor_admin_registration_status') == 'MO_2_FACTOR_CUSTOMER_REGISTERED_SUCCESS' && get_option( 'mo2f_miniorange_admin') != $current_user->ID){
-								if(get_user_meta($current_user->ID,'mo_2factor_user_registration_with_miniorange',true) != 'SUCCESS'){
-									mo2f_register_additional_admin($current_user);
-									delete_user_meta($current_user->ID,'mo_2factor_user_registration_status');
-								}
-							}
+							if($mo2f_active_tab == 'mobile_configure') {
+								instruction_for_mobile_registration($current_user);
+							}else if($mo2f_active_tab == 'mo2f_help'){
+								mo2f_show_help_and_troubleshooting($current_user);
+							}else if($mo2f_active_tab == 'mo2f_demo'){
+								show_2_factor_login_demo($current_user);
+							}else if(current_user_can( 'manage_options' ) && $mo2f_active_tab == 'mo2f_login'){
+								show_2_factor_login_settings($current_user);
+							}else{
 							
-							if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_OTP_DELIVERED_SUCCESS' || get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_OTP_DELIVERED_FAILURE'){
-								mo2f_show_otp_validation_page();
-							} else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_INITIALIZE_MOBILE_REGISTRATION'){
-								if(isset($_SESSION[ 'mo2f_qrCode' ])){
-									initialize_mobile_registration($current_user);
-								}else{
-									mo2f_get_qr_code_for_mobile(get_user_meta($current_user->ID,'mo_2factor_map_id_with_email',true),$current_user->ID);
-									initialize_mobile_registration($current_user);
+							
+								if(get_option( 'mo_2factor_admin_registration_status') == 'MO_2_FACTOR_CUSTOMER_REGISTERED_SUCCESS' && get_option( 'mo2f_miniorange_admin') != $current_user->ID){
+									if(get_user_meta($current_user->ID,'mo_2factor_user_registration_with_miniorange',true) != 'SUCCESS'){
+										mo2f_register_additional_admin($current_user);
+										delete_user_meta($current_user->ID,'mo_2factor_user_registration_status');
+									}
 								}
 								
-							} else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_PLUGIN_SETTINGS'){
-								show_2_factor_login_settings();
-							}else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_VERIFY_CUSTOMER') {
-								mo2f_show_verify_password_page();
-							} else if(!mo2f_is_customer_registered()){
-								delete_option('password_mismatch');
-								mo2f_show_new_registration_page($current_user);
-							} 
+								if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_OTP_DELIVERED_SUCCESS' || get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_OTP_DELIVERED_FAILURE'){
+									mo2f_show_otp_validation_page();
+								} else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_INITIALIZE_MOBILE_REGISTRATION'){
+									mo2f_show_instruction_to_allusers($current_user);
+								} else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_PLUGIN_SETTINGS'){
+									mo2f_show_instruction_to_allusers($current_user);
+									
+								}else if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) == 'MO_2_FACTOR_VERIFY_CUSTOMER') {
+									mo2f_show_verify_password_page();
+								} else if(!mo2f_is_customer_registered()){
+									delete_option('password_mismatch');
+									mo2f_show_new_registration_page($current_user);
+								} 
+							
+							}
 						?>
 					</td>
 					<td style="vertical-align:top;padding-left:1%;">
@@ -46,16 +66,16 @@
 		</div>
 		<?php
 	}
-
+	
 	function mo2f_show_new_registration_page($current_user) {
 		?>
 			<!--Register with miniOrange-->
 			<form name="f" method="post" action="">
 				<input type="hidden" name="option" value="mo_auth_register_customer" />
 				<div class="mo2f_table_layout">
-					<div id="toggle1" class="panel_toggle"><h3>Register with miniOrange</h3></div>
+					<h3>Register with miniOrange</h3><hr>
 					<div id="panel1">
-						<div><b>Please enter a valid email id that you have access to. You will be able to move forward after verifying an OTP that we will send to this email.</b></div><br />
+						<div><b>Please enter a valid email id that you have access to. You will be able to move forward after verifying an OTP that we will be sending to this email.</b></div><br />
 						<table class="mo2f_settings_table">
 							<tr>
 							<td><b><font color="#FF0000">*</font>Email:</b></td>
@@ -81,9 +101,11 @@
 							<tr><td>&nbsp;</td></tr>
 						  <tr>
 							<td>&nbsp;</td>
-							<td><input type="submit" name="submit" value="Next" class="button button-primary button-large" /></td>
+							<td><input type="submit" name="submit" value="Submit" class="button button-primary button-large" /></td>
 						  </tr>
 						</table>
+						<br>
+						
 					</div>
 				</div>
 			</form>
@@ -99,16 +121,17 @@
 		<!-- Enter otp -->
 		
 		<div class="mo2f_table_layout">
-			<div id="toggle1" class="panel_toggle"><h3>Validate OTP</h3></div>
+			<h3>Validate OTP</h3><hr>
 			<div id="panel1">
 				<table class="mo2f_settings_table">
 					<form name="f" method="post" id="mo_2f_otp_form" action="">
 						<input type="hidden" name="option" value="mo_2factor_validate_otp" />
 							<tr>
 								<td><b><font color="#FF0000">*</font>Enter OTP:</b></td>
-								<td colspan="2"><input class="mo2f_table_textbox" autofocus="true" type="text" name="otp_token" required placeholder="Enter OTP" style="width:61%;"/></td>
+								<td colspan="2"><input class="mo2f_table_textbox" autofocus="true" type="text" name="otp_token" required placeholder="Enter OTP" style="width:95%;"/></td>
+								<td><a href="#resendotplink">Resend OTP ?</a></td>
 							</tr>
-							<tr><td colspan="3"></td></tr>
+							
 							<tr>
 								<td>&nbsp;</td>
 								<td style="width:17%">
@@ -116,74 +139,73 @@
 
 						</form>
 						<form name="f" method="post" action="">
-							<td>
-							<input type="submit" name="mo_resend" id="mo_resend_otp" value="Resend OTP" class="button button-primary button-large button-green" />
-							<input type="hidden" name="option" value="mo_2factor_resend_otp"/>
-							</td>
-							</tr>
+						<td>
+						<input type="hidden" name="option" value="mo_2factor_gobackto_registration_page"/>
+							<input type="submit" name="mo2f_goback" id="mo2f_goback" value="Back" class="button button-primary button-large" /></td>
 						</form>
+						</td>
+						</tr>
+						<form name="f" method="post" action="" id="resend_otp_form">
+							<input type="hidden" name="option" value="mo_2factor_resend_otp"/>
+						</form>
+						
 				</table>
-				<form name="f" method="post" action="">
-					<input type="hidden" name="option" value="mo_2factor_gobackto_registration_page"/>
-					<br /><br /><br />
-					<div><center><input type="submit" name="mo2f_goback" id="mo2f_goback" value="Back" class="button button-primary button-large" /></center></div>
-					
-					
-				</form>
+				</div>
+				<div>	
+					<script>
+						jQuery('a[href="#resendotplink"]').click(function(e) {
+							jQuery('#resend_otp_form').submit();
+						});
+					</script>
+		
+			<br><br>
 			</div>
-			<br/>
-			<br/>
+			
+			
+						
 		</div>
 					
 	<?php
 	}
 	
-	function miniorange_2_factor_user_roles() {
-		global $wp_roles;
-		if (!isset($wp_roles))
-			$wp_roles = new WP_Roles();
+	function show_2_factor_login_settings($current_user) {
 		
-		print '<div>';
-		foreach($wp_roles->role_names as $id => $name) {	
-			$setting = get_option('mo2fa_'.$id);
-			$setting = $setting === false || $setting ? 1 : 0;
-			print '<input type="checkbox" name="mo2fa_'.$id.'" value="1" '.($setting ? 'checked="checked"' :'').'> <b>'.$name."</b><br>\n";
-		}
-		print '</div>';
-	}
-	function miniorange_2_factor_get_login_form_settings() {	
-		if(trim(get_option('mo_2f_login_type_enabled')) == 'mobile_only'){
-			print '<input type="radio" name="mo_2f_enabled" value="mobile_only" checked="checked"><b>Enable miniOrange 2-factor</b><br >';
-			print '<input type="radio" name="mo_2f_enabled" value="password_with_mobile" ><b>Enable miniOrange 2-factor with password</b><br>';
-		}else if(trim(get_option(mo_2f_login_type_enabled)) == 'password_with_mobile'){
-			print '<input type="radio" name="mo_2f_enabled" value="mobile_only" ><b>Enable miniOrange 2-factor</b><br>';
-			print '<input type="radio" name="mo_2f_enabled" value="password_with_mobile" checked="checked" ><b>Enable miniOrange 2-factor with password</b><br >';
-		}else{
-			print '<input type="radio" name="mo_2f_enabled" value="mobile_only" checked="checked" >'."<b>Enable miniOrange 2-factor</b><br>";
-			print '<input type="radio" name="mo_2f_enabled" value="password_with_mobile" >'."<b>Enable miniOrange 2-factor with password</b><br>";
-		}
-	}
-	
-	function show_2_factor_login_settings() {
-		echo mo2f_show_instruction_to_allusers();
 		?>
-		<br />
-		<div class="mo2f_small_layout">
+	<div class="mo2f_table_layout">
+			<?php echo mo2f_check_if_registered_with_miniorange($current_user); ?>
 			<form name="f" id="mowp_admin_activationform" method="post" action="">
-				<h3>Plugin activation settings</h3>
-				<input type="checkbox" id="mo2f_adminrole_activation" name="mo2f_adminrole_activation" value="1" <?php echo !get_option('mo2f_admin_disabled_status') ? 'checked="checked"' : ''; ?> />Enable plugin for admins.
+				<h3>Login Settings</h3>
+				<hr><br>
+				<input type="checkbox" id="mo2f_adminrole_activation" name="mo2f_adminrole_activation" value="1" <?php echo !get_option('mo2f_admin_disabled_status') ? 'checked="checked"' : ''; if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) != 'MO_2_FACTOR_PLUGIN_SETTINGS'){ echo 'disabled';} ?> />Enable 2-Factor for admins.
+				<br><br>
+				<div id="mo2f_note"><b>Note:</b> This option is checked by default. It will enable 2-Factor only for admins, other users can still login with their password.</div>
 				<input type="hidden" name="option" value="mo_auth_admin_activation" />
-			</form>
+			</form><br/>
 			<form name="f" id="mowp_activationform" method="post" action="">
-				<input type="checkbox" id="mo2f_role_activation" name="mo2f_role_activation" value="1" <?php echo get_option('mo2f_disabled_status') ? 'checked="checked"' : ''; ?> />Enable plugin for all other users.<br /><br />
-				<div id="mo2f_note"><b>Note:</b> It is recommended that you logout and test the login before you enable it for your users.</div>
+				<input type="checkbox" id="mo2f_role_activation" name="mo2f_role_activation" value="1" <?php echo get_option('mo2f_disabled_status') ? 'checked="checked"' : ''; if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) != 'MO_2_FACTOR_PLUGIN_SETTINGS'){ echo 'disabled';}?> />Enable 2-Factor for all other users.<br /><br />
+				<div id="mo2f_note"><b>Note:</b> Checking this option will enable 2-Factor for all users. Make sure you have tested login with 2-factor before enabling it for all users.</div>
 				<input type="hidden" name="option" value="mo_auth_user_activation" />
 			</form><br />
 			<form name="f" id="mowp_forgotphone_form" method="post" action="">
-				<input type="checkbox" id="mo2f_forgotphone" name="mo2f_forgotphone" value="1" <?php echo !get_option('mo2f_enable_forgotphone') ? 'checked="checked"' : ''; ?> />Enable Forgot Phone.<br /><br /><div id="mo2f_note"><b>Note:</b> Checking this option will enable Forgot My Phone for all the users during Login. An OTP over registered email will be send to verify the user. User has to enter OTP to bypass mobile authentication.</div>
+				<input type="checkbox" id="mo2f_forgotphone" name="mo2f_forgotphone" value="1" <?php echo !get_option('mo2f_enable_forgotphone') ? 'checked="checked"' : ''; if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) != 'MO_2_FACTOR_PLUGIN_SETTINGS'){ echo 'disabled';} ?> />Enable Forgot Phone.<br />
+				<br /><div id="mo2f_note"><b>Note:</b> This option is checked by default so that if anybody has forgotten his phone or phone is lost/stolen/discharged. An OTP over registered email will be send to verify the user. User has to enter OTP to bypass login with your phone.</div>
 				<input type="hidden" name="option" value="mo2f_forgotphone_activation" />
+			</form><br />
+			<form name="f" id="mowp_disable_defaultform" method="post" action="">
+				<input type="checkbox" id="mo2f_loginwith_phone" name="mo2f_loginwith_phone" value="1" <?php echo get_option('mo2f_show_loginwith_phone') ? 'checked="checked"' : ''; if(get_user_meta($current_user->ID,'mo_2factor_user_registration_status',true) != 'MO_2_FACTOR_PLUGIN_SETTINGS'){ echo 'disabled';} ?> />Show Login with phone only. (Hide default login form)&nbsp;&nbsp;
+				<a class="btn btn-link" data-toggle="collapse" href="#preview" aria-expanded="false">See preview</a><br>
+				<br><div id="mo2f_note"><b>Note:</b> Checking this option will hide the default login form. Click above link to see the preview. Users who have not setup 2-Factor can still login with their password.</div>
+				<input type="hidden" name="option" value="mo2f_loginphone_activation" /><br>
+				<div class="collapse" id="preview" style="height:300px;">
+				<center>
+				<img style="height:300px;" src="<?php echo plugins_url('includes/images/help/login-help-3.png', __FILE__ )?>" >
+				 </center>
+				 </div>
 			</form>
+			<br><br>
 		</div>
+			
+		
 		<script>
 			 jQuery('#mo2f_role_activation').change(function() {
 				jQuery('#mowp_activationform').submit();
@@ -193,6 +215,9 @@
 			});
 			jQuery('#mo2f_forgotphone').change(function() {
 				jQuery('#mowp_forgotphone_form').submit();
+			});
+			jQuery('#mo2f_loginwith_phone').change(function() {
+				jQuery('#mowp_disable_defaultform').submit();
 			});
 		</script>
 	<?php
@@ -204,7 +229,7 @@
 			<form name="f" method="post" action="">
 			<input type="hidden" name="option" value="mo_auth_verify_customer" />
 			<div class="mo2f_table_layout">
-			<div id="toggle1" class="panel_toggle"><h3>Login with miniOrange</h3></div>
+			<h3>Login with miniOrange</h3><hr>
 			<div id="panel1">
 			<p>Enter your miniOrange Email and Password.</p>
 			<br/>
@@ -220,11 +245,16 @@
 				<tr><td colspan="2">&nbsp;</td></tr>
 				<tr>
 				<td>&nbsp;</td>
-				<td><input type="submit" name="submit" class="button button-primary button-large" />&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<a target="_blank" href="<?php echo get_option('mo2f_host_name') . "/moas/idp/userforgotpassword"; ?>">Forgot your password?</a></td>
+				<td><input type="submit" name="submit" value="Submit" class="button button-primary button-large" />&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<a target="_blank" href="<?php echo get_option('mo2f_host_name') . "/moas/idp/userforgotpassword"; ?>">Forgot your password?</a></td>
 			  </tr>
 			</table>
+		
+			</div><br><br>
 			</div>
-			</div>
+			
+						
+			
+					
 			</form>
 	<?php	}
 	
@@ -232,15 +262,15 @@
 	?>
 		<form name="f" method="post" action="">
 			<div class="mo2f_table_layout">
-				<div id="toggle1" class="panel_toggle"><center><p style="font-size:17px;">miniOrange 2 Factor Authentication has been enabled for you. Please set up your account and register yourself by following the steps.</p></center></div>
+				<div><center><p style="font-size:17px;">miniOrange 2 Factor Authentication has been enabled for you. Please set up your account and register yourself by following the steps.</p></center></div>
 				<div id="panel1">
 					<table class="mo2f_settings_table">
 						
 						<tr>
-							<td><center><div class="alert-box"><input type="email" autofocus="true" name="mo_useremail" style="width: 70%;border: 2px solid #F7934D;text-align: center;height: 40px;font-size: 24px;border-radius:5px;" required placeholder="person@example.com" value="<?php echo $current_user->user_email;?>"/></div></center></td>
+							<td><center><div class="alert-box"><input type="email" autofocus="true" name="mo_useremail" style="width:48%;text-align: center;height: 40px;font-size:18px;border-radius:5px;" required placeholder="person@example.com" value="<?php echo $current_user->user_email;?>"/></div></center></td>
 						</tr>
 						<tr>
-							<td><center><p>Please enter a valid email id that you have access to. You will be able to login after verifying an OTP that we will send to this email in case you forgot or lost your phone.</p></center></td>
+							<td><center><p>Please enter a valid email id that you have access to. We need this email to send OTP in case you forgot or lost your phone.</p></center></td>
 						</tr>
 						<tr><td></td></tr>
 						<tr><td></td></tr>
